requirement.txt
numpy>=1.24
pandas>=2.0
scipy>=1.10
PyYAML>=6.0
yfinance>=0.2
arch>=6.0
pytest>=7.0
jupyterlab>=4.0
ipykernel>=6.0
matplotlib>=3.7
tqdm>=4.66